pyproject.toml
[tool.poetry]
name = "django-aeotrade-connector"
version = "0.1.0rc3"
description = "A connector development framework adapted to Django"
authors = ["hero <wanghao@aeotrade.com>"]
readme = "README.md"
packages = [{ include = "aeotrade_connector" }]

[tool.poetry.dependencies]
python = ">=3.10,<4"
aio-pika = "^9.5.3"
pydantic = "^2.10.3"
xmltodict = "^0.14.2"
httpx = "^0.28.1"
aiohttp = "^3.11.10"
requests = "^2.32.3"
django-apscheduler = "^0.7.0"
django-redis = "^5.4.0"
djangorestframework = "^3.15.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
pre-commit = "^4.0.1"
django = "^5.1.4"
types-redis = "^4.6.0.20241004"
pymysql = "^1.1.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[tool.mypy]
ignore_missing_imports = true
